[BREAKING] Refactor module usage, don't use .default
This stops using ES6 modules, and also moves the "main" for this package
to `index.js`, allowing direct import of the two blob classes.
