+++ package.json
@@ -38,5 +38,5 @@
     "taskcluster-client": "^3.1.1",
     "uuid": "^3.1.0"
   },
-  "main": "./lib/datacontainer"
+  "main": "./lib/index"
 }
