+++ src/index.js
@@ -1,0 +1,6 @@
+const {DataContainer} = require('./datacontainer');
+module.exports.DataContainer = DataContainer;
+
+const {DataBlockBlob, AppendDataBlob} = require('./datablob');
+module.exports.DataBlockBlob = DataBlockBlob;
+module.exports.AppendDataBlob = AppendDataBlob;
