+++ test/appendblob_test.js
@@ -1,10 +1,8 @@
-import assume        from 'assume';
-import DataContainer from '../lib/datacontainer';
-import {AppendDataBlob}    from '../lib/datablob';
-import uuid          from 'uuid';
-import _debug        from 'debug';
-const debug = _debug('azure-blob-storage-test:append-data-blob');
-import {logSchema, credentials}      from './helpers';
+const assume = require('assume');
+const debug = require('debug')('azure-blob-storage-test:data-blob');
+const {logSchema, credentials} = require('./helpers');
+const uuid = require('uuid');
+const {DataContainer, AppendDataBlob} = require('../lib');
 
 suite('Azure Blob Storage - Append Data Blob Tests', () => {
   let dataContainer;
