+++ test/version_test.js
@@ -1,10 +1,8 @@
-import _debug             from 'debug';
-const debug = _debug('azure-blob-storage-test:version');
-import {schemaV1, schemaV2, logSchema, logSchemaV2, credentials}      from './helpers';
-import uuid               from 'uuid';
-import DataContainer      from '../lib/datacontainer';
-import assume             from 'assume';
-import {DataBlockBlob}    from '../lib/datablob';
+const {schemaV1, schemaV2, logSchema, logSchemaV2, credentials} = require('./helpers');
+const assume = require('assume');
+const debug = require('debug')('azure-blob-storage-test:version');
+const uuid = require('uuid');
+const {DataContainer, DataBlockBlob} = require('../lib');
 
 describe('Azure Blob Storage - Data Container Version Support', () => {
